move wallet.save and load to files.go
