+++ sia/wallet/wallet.go
@@ -11,13 +11,6 @@ import (
 	"github.com/NebulousLabs/Sia/signatures"
 )
 
-// AddressKey is how we serialize and store spendable addresses on
-// disk.
-type AddressKey struct {
-	SpendConditions consensus.SpendConditions
-	SecretKey       signatures.SecretKey
-}
-
 // Wallet holds your coins, manages privacy, outputs, ect. The balance reported
 // ignores outputs you've already spent even if they haven't made it into the
 // blockchain yet.
@@ -61,61 +54,8 @@ func New(filename string) (w *Wallet, err error) {
 		transactions:       make(map[string]*openTransaction),
 	}
 
-	// Check if the file exists, then read it into memory.
-	if _, err = os.Stat(filename); os.IsNotExist(err) {
-		err = nil
-		return
-	}
-	contents, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return
-	}
+	w.Load(filename)
 
-	// Unmarshal the spendable addresses and put them into the wallet.
-	var keys []AddressKey
-	err = encoding.Unmarshal(contents, &keys)
-	if err != nil {
-		return
-	}
-	for _, key := range keys {
-		newSpendableAddress := &spendableAddress{
-			spendableOutputs: make(map[consensus.OutputID]*spendableOutput),
-			spendConditions:  key.SpendConditions,
-			secretKey:        key.SecretKey,
-		}
-		w.spendableAddresses[key.SpendConditions.CoinAddress()] = newSpendableAddress
-	}
-	return
-}
-
-// Reset implements the core.Wallet interface.
-func (w *Wallet) Reset() error {
-	w.Lock()
-	defer w.Unlock()
-	w.spentCounter++
-	return nil
-}
-
-// Save implements the core.Wallet interface.
-func (w *Wallet) Save() (err error) {
-	// Add every known spendable address + secret key.
-	var i int
-	keys := make([]AddressKey, len(w.spendableAddresses))
-	for _, spendableAddress := range w.spendableAddresses {
-		key := AddressKey{
-			SpendConditions: spendableAddress.spendConditions,
-			SecretKey:       spendableAddress.secretKey,
-		}
-		keys[i] = key
-		i++
-	}
-
-	//  write the file
-	fileData := encoding.Marshal(keys)
-	if err != nil {
-		return
-	}
-	err = ioutil.WriteFile(w.saveFilename, fileData, 0666)
 	return
 }
